Removed all xvba dependecies
xvba is cumbersome and induces a lot of errors when exporting/importing from excel.  At some point it even crashes and corrupted completely the file.

I decided to get rid of all the xvba folder and xvba extensions, and to use custom import/export module in the dev sheet of the linelist designer.

I strongly discourage using xvba after a few months of struggle betwen excel broken files and empty import files in vscodes.

Plus xvba is not opensource, I am completely unable to give feedbacks or to improve the extension without explicitly writing to the dev by mail.